+++ core/src/dht/dht_reducers.rs
@@ -8,12 +8,11 @@ use crate::{
 use holochain_core_types::{
     cas::content::{Address, AddressableContent},
     crud_status::{create_crud_link_eav, create_crud_status_eav, CrudStatus, STATUS_NAME},
-    eav::EntityAttributeValue,
+    eav::{EntityAttributeValueIndex, IndexQuery},
     entry::Entry,
     error::HolochainError,
 };
-
-use std::{collections::HashSet, convert::TryFrom, str::FromStr, sync::Arc};
+use std::{collections::BTreeSet, convert::TryFrom, str::FromStr, sync::Arc};
 
 // A function that might return a mutated DhtStore
 type DhtReducer = fn(Arc<Context>, &DhtStore, &ActionWrapper) -> Option<DhtStore>;
@@ -74,7 +73,7 @@ pub(crate) fn reduce_hold_entry(
         let meta_storage = &new_store.meta_storage().clone();
         create_crud_status_eav(&entry.address(), CrudStatus::Live)
             .map(|status_eav| {
-                let meta_res = (*meta_storage.write().unwrap()).add_eav(&status_eav);
+                let meta_res = (*meta_storage.write().unwrap()).add_eavi(&status_eav);
                 meta_res
                     .map(|_| Some(new_store))
                     .map_err(|err| {
@@ -119,11 +118,14 @@ pub(crate) fn reduce_add_link(
         );
         Some(new_store)
     } else {
-        let eav =
-            EntityAttributeValue::new(link.base(), &format!("link__{}", link.tag()), link.target());
+        let eav = EntityAttributeValueIndex::new(
+            link.base(),
+            &format!("link__{}", link.tag()),
+            link.target(),
+        );
         eav.map(|e| {
             let storage = new_store.meta_storage();
-            let result = storage.write().unwrap().add_eav(&e);
+            let result = storage.write().unwrap().add_eavi(&e);
             new_store
                 .actions_mut()
                 .insert(action_wrapper.clone(), result.map(|_| link.base().clone()));
@@ -150,7 +152,7 @@ pub(crate) fn reduce_update_entry(
     let closure_store = new_store.clone();
     let new_status_eav_option = create_crud_status_eav(latest_old_address, CrudStatus::Modified)
         .map(|new_status_eav| {
-            let res = (*meta_storage.write().unwrap()).add_eav(&new_status_eav);
+            let res = (*meta_storage.write().unwrap()).add_eavi(&new_status_eav);
             res.map(|_| None)
                 .map_err(|err| {
                     closure_store
@@ -170,7 +172,7 @@ pub(crate) fn reduce_update_entry(
         // Update crud-link
         create_crud_link_eav(latest_old_address, new_address)
             .map(|crud_link_eav| {
-                let res = (*meta_storage.write().unwrap()).add_eav(&crud_link_eav);
+                let res = (*meta_storage.write().unwrap()).add_eavi(&crud_link_eav);
                 let res_option = res.clone().ok();
                 res_option
                     .and_then(|_| {
@@ -236,10 +238,11 @@ fn reduce_remove_entry_inner(
     // pre-condition: Current status must be Live
     // get current status
     let meta_storage = &new_store.meta_storage().clone();
-    let maybe_status_eav = meta_storage.read().unwrap().fetch_eav(
+    let maybe_status_eav = meta_storage.read().unwrap().fetch_eavi(
         Some(latest_deleted_address.clone()),
         Some(STATUS_NAME.to_string()),
         None,
+        IndexQuery::default(),
     );
     if let Err(err) = maybe_status_eav {
         return Err(err);
@@ -249,9 +252,9 @@ fn reduce_remove_entry_inner(
     // TODO waiting for update/remove_eav() assert!(status_eavs.len() <= 1);
     // For now checks if crud-status other than Live are present
     let status_eavs = status_eavs
-        .iter()
+        .into_iter()
         .filter(|e| CrudStatus::from_str(String::from(e.value()).as_ref()) != Ok(CrudStatus::Live))
-        .collect::<HashSet<&EntityAttributeValue>>();
+        .collect::<BTreeSet<EntityAttributeValueIndex>>();
     if !status_eavs.is_empty() {
         return Err(HolochainError::ErrorGeneric(String::from(
             "entry_status != CrudStatus::Live",
@@ -266,14 +269,14 @@ fn reduce_remove_entry_inner(
     }
     let new_status_eav = result.expect("should unwrap eav");
     let meta_storage = &new_store.meta_storage().clone();
-    let res = (*meta_storage.write().unwrap()).add_eav(&new_status_eav);
+    let res = (*meta_storage.write().unwrap()).add_eavi(&new_status_eav);
     if let Err(err) = res {
         return Err(err);
     }
     // Update crud-link
     let crud_link_eav = create_crud_link_eav(latest_deleted_address, deletion_address)
         .map_err(|_| HolochainError::ErrorGeneric(String::from("Could not create eav")))?;
-    let res = (*meta_storage.write().unwrap()).add_eav(&crud_link_eav);
+    let res = (*meta_storage.write().unwrap()).add_eavi(&crud_link_eav);
     res.map(|_| latest_deleted_address.clone())
 }
 
@@ -302,6 +305,7 @@ pub mod tests {
     };
     use holochain_core_types::{
         cas::content::AddressableContent,
+        eav::IndexQuery,
         entry::{test_entry, test_sys_entry, Entry},
         link::Link,
     };
@@ -369,10 +373,12 @@ pub mod tests {
             new_dht_store = (*reduce(Arc::clone(&context), state.dht(), &action)).clone();
         }
         let storage = new_dht_store.meta_storage();
-        let fetched = storage
-            .read()
-            .unwrap()
-            .fetch_eav(Some(entry.address()), None, None);
+        let fetched = storage.read().unwrap().fetch_eavi(
+            Some(entry.address()),
+            None,
+            None,
+            IndexQuery::default(),
+        );
 
         assert!(fetched.is_ok());
         let hash_set = fetched.unwrap();
@@ -405,10 +411,12 @@ pub mod tests {
             new_dht_store = (*reduce(Arc::clone(&context), state.dht(), &action)).clone();
         }
         let storage = new_dht_store.meta_storage();
-        let fetched = storage
-            .read()
-            .unwrap()
-            .fetch_eav(Some(entry.address()), None, None);
+        let fetched = storage.read().unwrap().fetch_eavi(
+            Some(entry.address()),
+            None,
+            None,
+            IndexQuery::default(),
+        );
 
         assert!(fetched.is_ok());
         let hash_set = fetched.unwrap();