+++ dna_c_binding/src/lib.rs
@@ -253,17 +253,11 @@ pub unsafe extern "C" fn holochain_dna_get_capabilities_names(
 }
 
 #[cfg_attr(tarpaulin, skip)] //Tested in c_bindings_test by C based test code
-fn fn_names_as_vec(
-    dna: &Dna,
-    zome_name: &str,
-    capability_name: &str,
-) -> Option<Vec<*const c_char>> {
+fn fn_names_as_vec(dna: &Dna, zome_name: &str) -> Option<Vec<*const c_char>> {
     let result = dna
         .zomes
         .get(zome_name)?
-        .capabilities
-        .get(capability_name)?
-        .functions
+        .fn_declarations
         .iter()
         .map(|fn_declaration| {
             let raw = match CString::new(fn_declaration.name.clone()) {
@@ -281,15 +275,13 @@ fn fn_names_as_vec(
 pub unsafe extern "C" fn holochain_dna_get_function_names(
     ptr: *mut Dna,
     zome_name: *const c_char,
-    capability_name: *const c_char,
     string_vec: *mut CStringVec,
 ) {
     let dna = &*ptr;
 
     let zome_name = CStr::from_ptr(zome_name).to_string_lossy();
-    let capability_name = CStr::from_ptr(capability_name).to_string_lossy();
 
-    let fn_names = fn_names_as_vec(dna, &*zome_name, &*capability_name);
+    let fn_names = fn_names_as_vec(dna, &*zome_name);
     vec_char_to_cstringvec(fn_names, string_vec)
 }
 
@@ -297,17 +289,11 @@ pub unsafe extern "C" fn holochain_dna_get_function_names(
 fn fn_parameters_as_vec(
     dna: &Dna,
     zome_name: &str,
-    capability_name: &str,
     function_name: &str,
 ) -> Option<Vec<*const c_char>> {
     let result = dna
-        .zomes
-        .get(zome_name)?
-        .capabilities
-        .get(capability_name)?
-        .functions
-        .iter()
-        .find(|&function| function.name == function_name)?
+        .get_function_with_zome_name(zome_name, function_name)
+        .ok()?
         .inputs
         .iter()
         .map(|input| {
@@ -326,17 +312,15 @@ fn fn_parameters_as_vec(
 pub unsafe extern "C" fn holochain_dna_get_function_parameters(
     ptr: *mut Dna,
     zome_name: *const c_char,
-    capability_name: *const c_char,
     function_name: *const c_char,
     string_vec: *mut CStringVec,
 ) {
     let dna = &*ptr;
 
     let zome_name = CStr::from_ptr(zome_name).to_string_lossy();
-    let capability_name = CStr::from_ptr(capability_name).to_string_lossy();
     let function_name = CStr::from_ptr(function_name).to_string_lossy();
 
-    let fn_parameters = fn_parameters_as_vec(dna, &*zome_name, &*capability_name, &*function_name);
+    let fn_parameters = fn_parameters_as_vec(dna, &*zome_name, &*function_name);
     vec_char_to_cstringvec(fn_parameters, string_vec)
 }
 
@@ -401,11 +385,28 @@ mod tests {
                         "capabilities": {
                             "test capability": {
                                 "type": "public",
-                                "fn_declarations": [],
-                                "code": {
-                                    "code": ""
-                                }
+                                "functions": []
                             }
+                        },
+                        "fn_declarations": [
+                            {
+                                "name": "test",
+                                "inputs": [
+                                    {
+                                        "name": "post",
+                                        "type": "string"
+                                    }
+                                ],
+                                "outputs" : [
+                                    {
+                                        "name": "hash",
+                                        "type": "string"
+                                    }
+                                ]
+                            }
+                        ],
+                        "code": {
+                            "code": ""
                         },
                         "entry_types": {}
                     },
@@ -416,11 +417,28 @@ mod tests {
                         "capabilities": {
                             "test capability": {
                                 "type": "public",
-                                "fn_declarations": [],
-                                "code": {
-                                    "code": ""
-                                }
+                                "functions": []
                             }
+                        },
+                        "fn_declarations": [
+                            {
+                                "name": "test2",
+                                "inputs": [
+                                    {
+                                        "name": "address",
+                                        "type": "string"
+                                    }
+                                ],
+                                "outputs" : [
+                                    {
+                                        "name": "hash",
+                                        "type": "string"
+                                    }
+                                ]
+                            }
+                        ],
+                        "code": {
+                            "code": ""
                         },
                         "entry_types": {}
                     }